Refactor oauth.rs into modular components
Break down the large oauth.rs file into smaller modules:
- jwk: JSON Web Key functionality
- dpop: DPoP proof generation and verification
- token: OAuth token management
- session: OAuth session handling
- utils: Helper functions for OAuth operations

This modular structure improves code organization, maintainability,
and makes it easier to locate and update specific functionality.
