+++ server/src/oauth/mod.rs
@@ -1,0 +1,20 @@
+//! OAuth module for handling BlueskyOAuth operations
+//! This includes working with JWKs, DPoP proofs, tokens, and OAuth sessions
+
+// Re-export submodules
+pub mod db;
+pub mod dpop;
+pub mod jwk;
+pub mod session;
+pub mod token;
+pub mod utils;
+
+// Re-export main types and functions
+pub use dpop::*;
+pub use jwk::*;
+pub use session::*;
+pub use token::*;
+
+// Currently we only define database operations in their own file
+// The rest of the functionality is re-exported from mod.rs
+// This will be refactored into proper modules
