+++ server/src/oauth/token/mod.rs
@@ -1,141 +1,8 @@
-use base64ct::{Base64UrlUnpadded, Encoding};
-use cja::jobs::Job;
-use color_eyre::eyre::{eyre, WrapErr};
-use jsonwebtoken::Algorithm;
-use p256::{ecdsa::VerifyingKey, pkcs8::DecodePublicKey, EncodedPoint, PublicKey};
 use serde::{Deserialize, Serialize};
 use std::time::{SystemTime, UNIX_EPOCH};
+use color_eyre::eyre::{eyre, WrapErr};
 
-use crate::state::BlueskyOAuthConfig;
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct Jwk {
-    pub kty: String,
-    pub crv: String,
-    pub x: String,
-    pub y: String,
-    pub kid: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub alg: Option<String>,
-    #[serde(rename = "use", skip_serializing_if = "Option::is_none")]
-    pub use_: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub key_ops: Option<Vec<String>>,
-}
-
-/// Generate a JWK from a base64-encoded public key
-pub fn generate_jwk(public_key_base64: &str) -> cja::Result<Jwk> {
-    // Decode the base64-encoded public key
-    let decoded_key = base64::Engine::decode(
-        &base64::engine::general_purpose::STANDARD,
-        public_key_base64,
-    )
-    .wrap_err("Failed to decode base64-encoded public key")?;
-
-    // Log the first few bytes without revealing the whole key
-    let key_preview = if decoded_key.len() > 30 {
-        format!("{:?}...", &decoded_key[..30])
-    } else {
-        format!("{:?}", decoded_key)
-    };
-    tracing::debug!("Public key starts with: {}", key_preview);
-
-    // Convert the decoded key bytes to a string for PEM parsing
-    let key_str = std::str::from_utf8(&decoded_key)
-        .wrap_err("Failed to convert decoded public key to string")?;
-
-    // Parse the public key from PEM format
-    let verifying_key = VerifyingKey::from_public_key_pem(key_str)
-        .wrap_err_with(|| format!("Failed to parse public key. Key preview: {}", key_preview))?;
-
-    // Get the public key as an EncodedPoint
-    let public_key = PublicKey::from(verifying_key);
-    let encoded_point = EncodedPoint::from(public_key);
-
-    // Extract x and y coordinates
-    let x_bytes = encoded_point
-        .x()
-        .ok_or_else(|| eyre!("Failed to extract x coordinate"))
-        .wrap_err("Missing x coordinate in public key")?;
-    let y_bytes = encoded_point
-        .y()
-        .ok_or_else(|| eyre!("Failed to extract y coordinate"))
-        .wrap_err("Missing y coordinate in public key")?;
-
-    // Base64-URL encode the coordinates
-    let x = Base64UrlUnpadded::encode_string(x_bytes);
-    let y = Base64UrlUnpadded::encode_string(y_bytes);
-
-    // Generate a unique ID for the key
-    let key_id = generate_key_id(&x, &y)?;
-
-    Ok(Jwk {
-        kty: "EC".to_string(),
-        crv: "P-256".to_string(),
-        x,
-        y,
-        kid: key_id,
-        alg: Some("ES256".to_string()),
-        use_: Some("sig".to_string()),
-        key_ops: Some(vec!["verify".to_string()]),
-    })
-}
-
-/// Generate a key ID from the key's coordinates
-fn generate_key_id(x: &str, y: &str) -> cja::Result<String> {
-    use ring::digest::{Context, SHA256};
-
-    let mut context = Context::new(&SHA256);
-    context.update(x.as_bytes());
-    context.update(y.as_bytes());
-    let digest = context.finish();
-
-    Ok(Base64UrlUnpadded::encode_string(digest.as_ref()))
-}
-
-/// Calculate the JWK thumbprint for the given public key
-///
-/// This follows RFC 7638 for JWK Thumbprint calculation
-pub fn calculate_jwk_thumbprint(public_key_base64: &str) -> cja::Result<String> {
-    // First generate the JWK for the public key
-    let jwk = generate_jwk(public_key_base64)?;
-
-    // Create the canonical JWK representation with only the required fields in lexicographic order
-    let canonical_jwk = serde_json::json!({
-        "crv": jwk.crv,
-        "kty": jwk.kty,
-        "x": jwk.x,
-        "y": jwk.y
-    });
-
-    // Convert to a compact JSON string without whitespace
-    let canonical_json =
-        serde_json::to_string(&canonical_jwk).wrap_err("Failed to serialize canonical JWK")?;
-
-    // Calculate SHA-256 hash
-    use ring::digest::{digest, SHA256};
-    let digest = digest(&SHA256, canonical_json.as_bytes());
-
-    // Base64-URL encode the result
-    Ok(Base64UrlUnpadded::encode_string(digest.as_ref()))
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ClientMetadata {
-    pub client_id: String,
-    pub application_type: String,
-    pub grant_types: Vec<String>,
-    pub scope: Vec<String>,
-    pub response_types: Vec<String>,
-    pub redirect_uris: Vec<String>,
-    pub dpop_bound_access_tokens: bool,
-    pub token_endpoint_auth_method: String,
-    pub token_endpoint_auth_signing_alg: String,
-    pub jwks: Vec<Jwk>,
-    pub client_name: String,
-    pub client_uri: String,
-}
-
+/// JWT payload for token requests
 #[derive(Debug, Serialize)]
 pub struct TokenRequestPayload {
     pub iss: String,
@@ -146,16 +13,123 @@ pub struct TokenRequestPayload {
     pub iat: u64,
 }
 
+/// Response from the token endpoint
+#[derive(Debug, Deserialize, Clone)]
+pub struct TokenResponse {
+    pub access_token: String,
+    pub token_type: String,
+    pub expires_in: u64,
+    pub refresh_token: Option<String>,
+    pub scope: String,
+    /// DPoP confirmation key (JWK thumbprint)
+    #[serde(rename = "cnf")]
+    pub dpop_confirmation: Option<DPoPConfirmation>,
+}
+
+/// DPoP confirmation key in token response
+#[derive(Debug, Deserialize, Clone)]
+pub struct DPoPConfirmation {
+    /// JWK thumbprint
+    pub jkt: String,
+}
+
+/// Represents a complete set of OAuth tokens with metadata
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct OAuthTokenSet {
+    /// The access token for API requests
+    pub access_token: String,
+    /// The token type (usually "DPoP" for Bluesky)
+    pub token_type: String,
+    /// When the access token expires (as Unix timestamp)
+    pub expires_at: u64,
+    /// Refresh token for obtaining a new access token
+    pub refresh_token: Option<String>,
+    /// The scopes granted to this token
+    pub scope: String,
+    /// The Bluesky DID this token is associated with
+    pub did: String,
+    /// The Bluesky display name associated with this DID
+    pub display_name: Option<String>,
+    /// The Bluesky handle (username with domain) associated with this DID
+    pub handle: Option<String>,
+    /// DPoP confirmation key (JWK thumbprint)
+    pub dpop_jkt: Option<String>,
+    /// User ID that owns this token
+    pub user_id: Option<uuid::Uuid>,
+}
+
+impl OAuthTokenSet {
+    /// Create a new OAuthTokenSet from a TokenResponse
+    pub fn from_token_response(response: TokenResponse, did: String) -> Self {
+        let now = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap_or_default()
+            .as_secs();
+
+        Self {
+            access_token: response.access_token,
+            token_type: response.token_type,
+            expires_at: now + response.expires_in,
+            refresh_token: response.refresh_token,
+            scope: response.scope,
+            did,
+            display_name: None, // Will be updated later when we fetch profile data
+            handle: None,       // Will be updated later when we fetch profile data
+            dpop_jkt: response.dpop_confirmation.map(|cnf| cnf.jkt),
+            user_id: None,
+        }
+    }
+
+    /// Copy the display name and handle from another token
+    #[allow(dead_code)]
+    pub fn with_display_name_from(mut self, other: &OAuthTokenSet) -> Self {
+        self.display_name = other.display_name.clone();
+        self.handle = other.handle.clone();
+        self
+    }
+
+    /// Create a new OAuthTokenSet from a TokenResponse with a calculated JWK thumbprint
+    pub fn from_token_response_with_jwk(
+        response: &TokenResponse,
+        did: String,
+        public_key: &str,
+    ) -> cja::Result<Self> {
+        let mut token_set = Self::from_token_response(response.clone(), did);
+
+        // If there's no JWK thumbprint in the response, calculate it
+        if token_set.dpop_jkt.is_none() {
+            let calculated_jkt = crate::oauth::jwk::calculate_jwk_thumbprint(public_key)?;
+            token_set.dpop_jkt = Some(calculated_jkt);
+        }
+
+        Ok(token_set)
+    }
+
+    /// Check if the access token is expired
+    pub fn is_expired(&self) -> bool {
+        let now = SystemTime::now()
+            .duration_since(UNIX_EPOCH)
+            .unwrap_or_default()
+            .as_secs();
+
+        // Consider tokens expired 30 seconds before actual expiration
+        // to account for clock skew and network latency
+        self.expires_at < now + 30
+    }
+}
+
 /// Create a client assertion JWT for OAuth token requests
 pub fn create_client_assertion(
-    oauth_config: &BlueskyOAuthConfig,
+    oauth_config: &crate::state::BlueskyOAuthConfig,
     token_endpoint: &str,
     client_id: &str,
 ) -> cja::Result<String> {
     use std::io::Write;
     use std::process::Command;
-
     use tempfile::NamedTempFile;
+    use crate::oauth::utils::base64_url_encode;
+    use crate::oauth::utils::der_signature_to_raw_signature;
+    use jsonwebtoken::Algorithm;
 
     // Debug info
     let key_preview = if oauth_config.private_key.len() > 10 {
@@ -178,7 +152,7 @@ pub fn create_client_assertion(
         .as_secs();
 
     // Get the JWK for header
-    let jwk = generate_jwk(&oauth_config.public_key)?;
+    let jwk = crate::oauth::jwk::generate_jwk(&oauth_config.public_key)?;
 
     // Create the JWT header with the key ID
     let mut header = jsonwebtoken::Header::new(Algorithm::ES256);
@@ -301,365 +275,9 @@ pub fn create_client_assertion(
     Ok(token)
 }
 
-/// Helper function to create URL-safe base64 encoding without padding
-fn base64_url_encode(input: &[u8]) -> String {
-    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
-    URL_SAFE_NO_PAD.encode(input)
-}
-
-/// Convert a DER-encoded ECDSA signature to the raw format required for JWT
-/// For ES256, the signature must be a 64-byte array containing R and S values (each 32 bytes)
-fn der_signature_to_raw_signature(der_signature: &[u8]) -> cja::Result<Vec<u8>> {
-    use simple_asn1::{from_der, ASN1Block};
-
-    use color_eyre::eyre::eyre;
-
-    // Parse the DER-encoded signature
-    let blocks = from_der(der_signature).wrap_err("Failed to parse DER signature")?;
-
-    // The DER signature should be a SEQUENCE with two INTEGERs (r and s)
-    if blocks.len() != 1 {
-        return Err(eyre!("Expected 1 ASN.1 block, found {}", blocks.len()));
-    }
-
-    // Extract the r and s values from the sequence
-    let (r, s) = match &blocks[0] {
-        ASN1Block::Sequence(_, items) => {
-            if items.len() != 2 {
-                return Err(eyre!("Expected 2 items in sequence, found {}", items.len()));
-            }
-
-            let r = match &items[0] {
-                ASN1Block::Integer(_, r) => r,
-                _ => return Err(eyre!("Expected INTEGER for r value, found {:?}", items[0])),
-            };
-
-            let s = match &items[1] {
-                ASN1Block::Integer(_, s) => s,
-                _ => return Err(eyre!("Expected INTEGER for s value, found {:?}", items[1])),
-            };
-
-            (r, s)
-        }
-        _ => return Err(eyre!("Expected SEQUENCE, found {:?}", blocks[0])),
-    };
-
-    // Convert BigInt to bytes
-    let (_, r_bytes) = r.to_bytes_be(); // to_bytes_be returns (sign, bytes)
-    let (_, s_bytes) = s.to_bytes_be();
-
-    // Ensure R and S are each 32 bytes
-    let mut result = vec![0u8; 64];
-
-    // Fill result with R value (left-padded with zeros if needed)
-    let r_offset = 32 - r_bytes.len();
-    if r_offset > 0 {
-        // R is shorter than 32 bytes, need padding
-        result[r_offset..32].copy_from_slice(&r_bytes);
-    } else if r_bytes.len() > 32 {
-        // R is longer than 32 bytes, need truncation (shouldn't happen)
-        result[..32].copy_from_slice(&r_bytes[r_bytes.len() - 32..]);
-    } else {
-        // R is exactly 32 bytes
-        result[..32].copy_from_slice(&r_bytes);
-    }
-
-    // Fill result with S value (left-padded with zeros if needed)
-    let s_offset = 32 - s_bytes.len();
-    if s_offset > 0 {
-        // S is shorter than 32 bytes, need padding
-        result[32 + s_offset..].copy_from_slice(&s_bytes);
-    } else if s_bytes.len() > 32 {
-        // S is longer than 32 bytes, need truncation (shouldn't happen)
-        result[32..].copy_from_slice(&s_bytes[s_bytes.len() - 32..]);
-    } else {
-        // S is exactly 32 bytes
-        result[32..].copy_from_slice(&s_bytes);
-    }
-
-    Ok(result)
-}
-
-#[derive(Debug, Deserialize, Clone)]
-pub struct TokenResponse {
-    pub access_token: String,
-    pub token_type: String,
-    pub expires_in: u64,
-    pub refresh_token: Option<String>,
-    pub scope: String,
-    /// DPoP confirmation key (JWK thumbprint)
-    #[serde(rename = "cnf")]
-    pub dpop_confirmation: Option<DPoPConfirmation>,
-}
-
-/// DPoP confirmation key in token response
-#[derive(Debug, Deserialize, Clone)]
-pub struct DPoPConfirmation {
-    /// JWK thumbprint
-    pub jkt: String,
-}
-
-/// Represents a complete set of OAuth tokens with metadata
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OAuthTokenSet {
-    /// The access token for API requests
-    pub access_token: String,
-    /// The token type (usually "DPoP" for Bluesky)
-    pub token_type: String,
-    /// When the access token expires (as Unix timestamp)
-    pub expires_at: u64,
-    /// Refresh token for obtaining a new access token
-    pub refresh_token: Option<String>,
-    /// The scopes granted to this token
-    pub scope: String,
-    /// The Bluesky DID this token is associated with
-    pub did: String,
-    /// The Bluesky display name associated with this DID
-    pub display_name: Option<String>,
-    /// The Bluesky handle (username with domain) associated with this DID
-    pub handle: Option<String>,
-    /// DPoP confirmation key (JWK thumbprint)
-    pub dpop_jkt: Option<String>,
-    /// User ID that owns this token
-    pub user_id: Option<uuid::Uuid>,
-}
-
-impl OAuthTokenSet {
-    /// Create a new OAuthTokenSet from a TokenResponse
-    pub fn from_token_response(response: TokenResponse, did: String) -> Self {
-        let now = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-
-        Self {
-            access_token: response.access_token,
-            token_type: response.token_type,
-            expires_at: now + response.expires_in,
-            refresh_token: response.refresh_token,
-            scope: response.scope,
-            did,
-            display_name: None, // Will be updated later when we fetch profile data
-            handle: None,       // Will be updated later when we fetch profile data
-            dpop_jkt: response.dpop_confirmation.map(|cnf| cnf.jkt),
-            user_id: None,
-        }
-    }
-
-    // Function removed - not used in codebase
-
-    /// Copy the display name and handle from another token
-    #[allow(dead_code)]
-    pub fn with_display_name_from(mut self, other: &OAuthTokenSet) -> Self {
-        self.display_name = other.display_name.clone();
-        self.handle = other.handle.clone();
-        self
-    }
-
-    // Removed unused methods
-
-    /// Create a new OAuthTokenSet from a TokenResponse with a calculated JWK thumbprint
-    pub fn from_token_response_with_jwk(
-        response: &TokenResponse,
-        did: String,
-        public_key: &str,
-    ) -> cja::Result<Self> {
-        let mut token_set = Self::from_token_response(response.clone(), did);
-
-        // If there's no JWK thumbprint in the response, calculate it
-        if token_set.dpop_jkt.is_none() {
-            let calculated_jkt = calculate_jwk_thumbprint(public_key)?;
-            token_set.dpop_jkt = Some(calculated_jkt);
-        }
-
-        Ok(token_set)
-    }
-
-    /// Check if the access token is expired
-    pub fn is_expired(&self) -> bool {
-        let now = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-
-        // Consider tokens expired 30 seconds before actual expiration
-        // to account for clock skew and network latency
-        self.expires_at < now + 30
-    }
-}
-
-/// Create a DPoP (Demonstrating Proof-of-Possession) proof JWT for API requests
-/// This is required by Bluesky's OAuth implementation
-pub fn create_dpop_proof(
-    oauth_config: &BlueskyOAuthConfig,
-    http_method: &str,
-    endpoint_url: &str,
-    server_nonce: Option<&str>,
-) -> cja::Result<String> {
-    // Call the implementation with access token parameter as None
-    create_dpop_proof_impl(oauth_config, http_method, endpoint_url, server_nonce, None)
-}
-
-/// Create a DPoP proof that includes the access token hash (ath) claim
-/// This is needed for some PDS servers that require the ath claim
-pub fn create_dpop_proof_with_ath(
-    oauth_config: &BlueskyOAuthConfig,
-    http_method: &str,
-    endpoint_url: &str,
-    server_nonce: Option<&str>,
-    access_token: &str,
-) -> cja::Result<String> {
-    // Call the implementation with the access token
-    create_dpop_proof_impl(
-        oauth_config,
-        http_method,
-        endpoint_url,
-        server_nonce,
-        Some(access_token),
-    )
-}
-
-/// Internal implementation for creating DPoP proofs with or without access token hash
-fn create_dpop_proof_impl(
-    oauth_config: &BlueskyOAuthConfig,
-    http_method: &str,
-    endpoint_url: &str,
-    server_nonce: Option<&str>,
-    access_token: Option<&str>,
-) -> cja::Result<String> {
-    use base64::{engine::general_purpose::URL_SAFE_NO_PAD, Engine};
-    use sha2::{Digest, Sha256};
-    use std::io::Write;
-    use std::process::Command;
-    use std::time::{SystemTime, UNIX_EPOCH};
-    use tempfile::NamedTempFile;
-
-    let now = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .map_err(|e| eyre!("Failed to get current time: {}", e))?
-        .as_secs();
-
-    // Generate JWK for header
-    let jwk = generate_jwk(&oauth_config.public_key)?;
-
-    // Create the JWT header with the key ID
-    let header_json = serde_json::json!({
-        "alg": "ES256",
-        "typ": "dpop+jwt",
-        "jwk": {
-            "kty": jwk.kty,
-            "crv": jwk.crv,
-            "x": jwk.x,
-            "y": jwk.y,
-            "kid": jwk.kid
-        }
-    })
-    .to_string();
-
-    // Create the DPoP payload according to the spec
-    let mut payload_json = serde_json::json!({
-        "jti": uuid::Uuid::new_v4().to_string(),
-        "htm": http_method,
-        "htu": endpoint_url,
-        "iat": now,
-        "exp": now + 300, // 5 minutes in the future
-    });
-
-    // Add nonce claim if provided
-    if let Some(nonce) = server_nonce {
-        payload_json["nonce"] = nonce.to_owned().into();
-    }
-
-    // Add ath (access token hash) claim if access token is provided
-    if let Some(token) = access_token {
-        // Calculate SHA-256 hash of the access token
-        let mut hasher = Sha256::new();
-        hasher.update(token.as_bytes());
-        let token_hash = hasher.finalize();
-
-        // Base64url encode the hash
-        let ath = URL_SAFE_NO_PAD.encode(token_hash);
-
-        // Add the ath claim to the payload
-        payload_json["ath"] = ath.into();
-
-        tracing::debug!("Added access token hash (ath) to DPoP proof");
-    }
-
-    let payload_json = payload_json.to_string();
-
-    // Log the DPoP JWT components
-    tracing::debug!("DPoP Header: {}", header_json);
-    tracing::debug!("DPoP Payload: {}", payload_json);
-
-    // 1. Create Base64URL-encoded header
-    let header_b64 = base64_url_encode(header_json.as_bytes());
-
-    // 2. Create Base64URL-encoded payload
-    let payload_b64 = base64_url_encode(payload_json.as_bytes());
-
-    // 3. Combine to form the message to sign
-    let message = format!("{}.{}", header_b64, payload_b64);
-
-    // 4. Create a temporary file for the message
-    let mut message_file = NamedTempFile::new()
-        .map_err(|e| eyre!("Failed to create temporary file for message: {}", e))?;
-    message_file
-        .write_all(message.as_bytes())
-        .map_err(|e| eyre!("Failed to write message to temporary file: {}", e))?;
-
-    // 5. Create a temporary file for the ES256 private key
-    let mut key_file = NamedTempFile::new()
-        .map_err(|e| eyre!("Failed to create temporary file for private key: {}", e))?;
-
-    // Decode base64-encoded private key
-    let decoded_key = base64::Engine::decode(
-        &base64::engine::general_purpose::STANDARD,
-        &oauth_config.private_key,
-    )
-    .wrap_err("Failed to decode base64-encoded private key")?;
-
-    // Write the decoded PEM-encoded private key to the file
-    key_file
-        .write_all(&decoded_key)
-        .wrap_err("Failed to write private key to temporary file")?;
-
-    // 6. Use OpenSSL to create the signature
-    let output = Command::new("openssl")
-        .arg("dgst")
-        .arg("-sha256")
-        .arg("-sign")
-        .arg(key_file.path())
-        .arg("-binary")
-        .arg(message_file.path())
-        .output()
-        .wrap_err("Failed to execute OpenSSL for signing")?;
-
-    if !output.status.success() {
-        let error = String::from_utf8_lossy(&output.stderr);
-        return Err(eyre!("OpenSSL signing failed: {}", error));
-    }
-
-    // The raw signature from OpenSSL is in ASN.1 DER format
-    // We need to convert this to the R+S concatenated format for ES256 JWT
-    let signature_der = output.stdout;
-
-    // 7. Convert DER-encoded signature to raw R||S format for JWT
-    let signature_raw = der_signature_to_raw_signature(&signature_der)
-        .wrap_err("Failed to convert DER signature to raw format")?;
-
-    // 8. Base64URL-encode the raw signature
-    let signature_b64 = base64_url_encode(&signature_raw);
-
-    // 9. Combine to form the complete JWT
-    let token = format!("{}.{}.{}", header_b64, payload_b64, signature_b64);
-
-    Ok(token)
-}
-
 /// Exchange authorization code for access token
 pub async fn exchange_code_for_token(
-    oauth_config: &BlueskyOAuthConfig,
+    oauth_config: &crate::state::BlueskyOAuthConfig,
     token_endpoint: &str,
     client_id: &str,
     code: &str,
@@ -688,7 +306,7 @@ pub async fn exchange_code_for_token(
 
     while retries > 0 {
         // Create a fresh DPoP proof for each attempt, using the nonce from the previous response if available
-        let dpop_proof = create_dpop_proof(
+        let dpop_proof = crate::oauth::dpop::create_dpop_proof(
             oauth_config,
             "POST",
             token_endpoint,
@@ -820,7 +438,7 @@ pub async fn exchange_code_for_token(
 
 /// Refresh an OAuth token using the refresh token
 pub async fn refresh_token(
-    oauth_config: &BlueskyOAuthConfig,
+    oauth_config: &crate::state::BlueskyOAuthConfig,
     token_endpoint: &str,
     client_id: &str,
     refresh_token: &str,
@@ -845,7 +463,7 @@ pub async fn refresh_token(
 
     while retries > 0 {
         // Create a fresh DPoP proof for each attempt, using the nonce from the previous response if available
-        let dpop_proof = create_dpop_proof(
+        let dpop_proof = crate::oauth::dpop::create_dpop_proof(
             oauth_config,
             "POST",
             token_endpoint,
@@ -971,91 +589,6 @@ pub async fn refresh_token(
     Err(last_error.unwrap_or_else(|| eyre!("Token refresh failed after retries")))
 }
 
-/// Represents the data stored in a session during the OAuth flow
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct OAuthSession {
-    /// The user's DID
-    pub did: String,
-    /// Original redirect URI provided to the authorize endpoint
-    // pub redirect_uri: Option<String>,
-    /// State parameter passed to the authorize endpoint
-    pub state: Option<String>,
-    /// The authorization server's token endpoint
-    pub token_endpoint: String,
-    /// The timestamp when this session was created
-    pub created_at: u64,
-    /// Optional token set if the OAuth flow has completed
-    pub token_set: Option<OAuthTokenSet>,
-    /// PKCE code verifier - the original random string
-    pub code_verifier: Option<String>,
-    /// PKCE code challenge - the hashed and encoded verifier
-    pub code_challenge: Option<String>,
-    /// DPoP nonce from the server
-    pub dpop_nonce: Option<String>,
-}
-
-impl OAuthSession {
-    /// Create a new OAuth session with PKCE
-    pub fn new(did: String, state: Option<String>, token_endpoint: String) -> Self {
-        let now = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-
-        // Generate PKCE code verifier and challenge
-        let (code_verifier, code_challenge) = Self::generate_pkce_codes().unwrap_or({
-            // Fallback to empty strings if generation fails
-            (None, None)
-        });
-
-        Self {
-            did,
-            state,
-            token_endpoint,
-            created_at: now,
-            token_set: None,
-            code_verifier,
-            code_challenge,
-            dpop_nonce: None,
-        }
-    }
-
-    /// Generate PKCE code verifier and challenge
-    fn generate_pkce_codes() -> cja::Result<(Option<String>, Option<String>)> {
-        use rand::{thread_rng, RngCore};
-        use sha2::{Digest, Sha256};
-
-        // Generate a random code verifier (between 43 and 128 characters)
-        let mut code_verifier_bytes = [0u8; 64]; // 64 bytes = 128 chars in hex
-        thread_rng().fill_bytes(&mut code_verifier_bytes);
-
-        // Base64-URL encode the verifier
-        let code_verifier = Base64UrlUnpadded::encode_string(&code_verifier_bytes);
-
-        // Create code challenge using the S256 method
-        let mut hasher = Sha256::new();
-        hasher.update(code_verifier.as_bytes());
-        let hashed_verifier = hasher.finalize();
-        let code_challenge = Base64UrlUnpadded::encode_string(&hashed_verifier);
-
-        Ok((Some(code_verifier), Some(code_challenge)))
-    }
-
-    /// Check if this session is expired (older than 1 hour)
-    pub fn is_expired(&self) -> bool {
-        let now = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_secs();
-
-        // Sessions expire after 1 hour
-        self.created_at + 3600 < now
-    }
-}
-
-/// Database functions for managing OAuth sessions and tokens
-pub mod db;
-
 /// Resolve the token endpoint for a DID
 pub async fn resolve_token_endpoint_for_did(
     did: &str,
@@ -1123,7 +656,6 @@ pub async fn refresh_token_if_needed(
     state: &crate::state::AppState,
     token_endpoint: &str,
 ) -> cja::Result<Option<OAuthTokenSet>> {
-    use color_eyre::eyre::eyre;
     use tracing::{error, info};
 
     // Only refresh if token is expired and we have a refresh token
@@ -1135,7 +667,7 @@ pub async fn refresh_token_if_needed(
     let client_id = state.client_id();
 
     // Try to get the latest DPoP nonce
-    let dpop_nonce = match db::get_latest_nonce(state, &token.did).await {
+    let dpop_nonce = match crate::oauth::db::get_latest_nonce(state, &token.did).await {
         Ok(nonce) => nonce,
         Err(err) => {
             error!("Failed to get DPoP nonce: {:?}", err);
@@ -1181,12 +713,13 @@ pub async fn refresh_token_if_needed(
             };
 
             // Store the new token with encryption
-            if let Err(err) = db::store_token(state, &new_token).await {
+            if let Err(err) = crate::oauth::db::store_token(state, &new_token).await {
                 error!("Failed to store refreshed token: {:?}", err);
                 return Err(eyre!("Failed to store refreshed token: {:?}", err));
             }
 
             // Also fetch profile to update display name if needed
+            use cja::jobs::Job;
             if let Err(err) = crate::jobs::UpdateProfileInfoJob::from_token(&new_token)
                 .enqueue(state.clone(), "token_refresh".to_string())
                 .await
@@ -1217,11 +750,10 @@ pub async fn get_valid_token_by_did(
     did: &str,
     state: &crate::state::AppState,
 ) -> cja::Result<OAuthTokenSet> {
-    use color_eyre::eyre::eyre;
     use tracing::{error, info};
 
     // Get the token from the database with decryption
-    let token = match db::get_token(state, did).await? {
+    let token = match crate::oauth::db::get_token(state, did).await? {
         Some(token) => token,
         None => return Err(eyre!("No token found for DID: {}", did)),
     };
